Do not uninstall needed virtual dependencies
Virtual dependencies are uninstalled if all dependencies to the
component is removed. Virtual dependencies were calculated using recent
components in repository, which is wrong as the dependencies may have
changed in the repository, causing the dependencies to be uninstalled
although those are still needed. Fixed so that the uninstallable virtual
components are calculated from repository components if the component is
going to be updated, otherwise the dependeny is calculated from the
local installed packages.

Task-number: QTIFW-2573
Change-Id: If9cc59fa7f453d502ec1883f27273ef604128a6e
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>