Revert "Simplify QInstaller::generateTemporaryFileName()"
This reverts commit 270c5dfd9b6c86bec8eaf34d0a2c3d3b945ed5e5.

Reason for revert: This caused a regression when an absolute path pointing to a directory non-writable by current user was used.
QTemporaryFile uses its own file engine so we need to use QFile
instead to forward the calls to the elevated privilege installer
process with the Qt Installer Framework's RemoteFileEngine.

Task-number: QTIFW-2575
Change-Id: I47f5bea06f9e1689ee322d972b405d8127967a98
Reviewed-by: Katja Marttila <katja.marttila@qt.io>