Use archive sizes for sorting and calculating progress of Extract ops
This makes the progress appear smoother when the archives to extract
differ in size.

The sorting based on the archive sizes provides some improvement to
optimal thread utilization, as the larger archives get more time to
complete from the start. We are less likely to find ourselves in a
situation where any of the last operations left to be run with less
than ideal thread count take significantly more time to complete than
the others.

Task-number: QTIFW-2431
Change-Id: I2534b6f18997d73c4ce1fb8b26900e30e1c6a76f
Reviewed-by: Katja Marttila <katja.marttila@qt.io>