Speed up Next button click in tree view
Components were recalculated when leaving the tree view. This is
unnecessary as we have just calculated the components. Moving the
recalculation to Introduction page so alternate pressing of 'Add or
remove components' and 'Update components' will cause the recalculation
as we need that to keep the component install states in sync.

Task-number: QTIFW-2598
Task-number: QTIFW-694
Change-Id: Ia84a01b4f1d0b39b4844c02470b764c4940e943c
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>