Fix rollback of extract when archive lists file entries first
At least with the 7z format packaged with a recent archivegen version,
the directory entries from archives get processed last when extracting.

We remember the path of each extracted entry in the order the writing
happens, to be able to undo the extract later.

The directories need to be empty for the removing to succeed, so we
cannot rely that the entry order from extraction is correct. Ensure
files are removed before directories and that the directories are
removed in an order starting from the deepest subdirectory level.

Task-number: QTIFW-2594
Change-Id: I8a914c4bf69147b845b12f0481c49c93408b51fd
Reviewed-by: Katja Marttila <katja.marttila@qt.io>