AspectRatioLabel: prevent setting non-scaled pixmaps
The function signature for setting a pixmap for the label changed
from 'setPixmap(const QPixmap &pixmap)' to
'setPixmapAndUrl(const QPixmap &pixmap, const QString &url)' when
we enabled support for click URLs.

Redefine (again) the base class QLabel::setPixmap() slot to make
it harder for callers to set non-scaled pixmaps accidentally. This
is still a bit bad however as it is possible to get different
client-visible results by calling the method with a base class
pointer.

Task-number: QTIFW-2612
Change-Id: Ic79b3d71ffd5baf2108768ad3b867ddbcb33833b
Reviewed-by: Katja Marttila <katja.marttila@qt.io>