Print uninstallation progress on archive level with normal verbosity
Don't print the name of every single removed file and directory unless
the verbosity level is set to detailed, as this really chokes the UI
thread.

QPlainTextEdit would have had better performance for a real-time output
log widget than QTextEdit, but we cannot use it as we require support
for displaying and automatically aligning text also for RTL languages.

Task-number: QTIFW-2615
Change-Id: Ie2938941b32bb387c0528b894cd058f395561814
Reviewed-by: Katja Marttila <katja.marttila@qt.io>