Do not recalculate local installed components
No need to recalculate local installed components if no changes has been
made. This makes the component selection quicker if there are lots
of components to be installed.

Local components are already calculated and listed in LocalPackageHub.
Using this class directly instead of calculating own local component
hash table. This changes the localInstalledPackages() from QHash to QMap,
which is not as optimized for searching components, but overall performance
is still now way better.

Task-number: QTIFW-2522
Change-Id: I8c16060f4b4014f740be0c7c43ece5e659d92987
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>