Do not emit dataChanged when model check state changes
Single items check state is not tracked anywhere else but in the model,
so we do not need to send dataChanged -signal to every item when the
model check state changes. The correct check state is already set to
every item above, so this emit seems unnecessary. This has significant
effect on 'Select All', 'Deselect All' and 'Default' -buttons. If there
are hundreds of items in the tree, this speeds up the button press in
best case over 50 percent.

Task-number: QTIFW-2623
Change-Id: If3c2aaebcc0dd6619179c31a51c8a57fdcec5c66
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>