Workaround QSettings / QDateTime incompatibility with Qt 6 applications
This causes problems when installers created with IFW share config files
with Qt 6 applications. If the config file contains keys with QDateTime
values written by the other application, the installer and maintenance
tool will segfault when attempting to sync the file, even if we do not
overwrite the values.

This is fixed to recent Qt 5 and 6 versions with the submitted patches
to QTBUG-102334. But as this requires updating the baseline Qt version
for both the Qt IFW and the Qt 6 application, apply the workaround
suggested in the issue comments, to avoid the crash in meantime.

Task-number: QTIFW-2610
Change-Id: I968e293324775dc650315e223f549a81657a27a3
Reviewed-by: Katja Marttila <katja.marttila@qt.io>