Adjust the 'ready to install' message to avoid repeating the app name
The text was using just the word 'Setup' which is inconsistent with other
places which usually refer to the application as '<product> Setup'. Adding
the product name would've been possible, but it seemed unnecessary to
repeat the installer name again right at the last step. So instead replace
that with a general message explaining that all needed information has been
collected.

Task-Name: SQUISH-9672
Change-Id: Ia253f4f921431f272bc1d69352a5fa7179e84d0c
Reviewed-by: Katja Marttila <katja.marttila@qt.io>