Swap priority of remote and binary resource package sources
When a component with the same name and equal version is available from
both remote repository and installer binary, it makes more sense to
use the installer resource as the package source for the component.

This also fixes the issue of erroneously registering duplicate archives
to a single resource collection. Because both the internal resources and
downloaded archives are registered in the same naming scheme of
"installer://<component>/<version><archive>", we would register the
same archive second time when using the remote source. If the packager
has not explicitly added Extract operations for the component, these
would be created automatically by looking for archive entries from the
component's resource collection, and so the same operation would be
added twice.

Task-number: QTIFW-2242
Change-Id: I82d4ed8a94a67d982b016bf7cec972e43bef8f03
Reviewed-by: Katja Marttila <katja.marttila@qt.io>