Match title and message text on the first wizard page
The title of the page was replicating that product name and the word
'Setup', but was not doing so consistently with either the message text
or the title. Since the titles of the other pages do not refer the product
and the product is mentioned in the message too, lets make this a short
'Welcome' instead.

Since this is vastly different from the original text, I decided to adjust
all translations for this string to unfinished - except for the german one
which I can easily contribute myself.

Task-Name: SQUISH-9672
Change-Id: Ia410d6e256b59b858f54082d17513a0b58a18546
Reviewed-by: Katja Marttila <katja.marttila@qt.io>
Reviewed-by: Tino Pyssysalo <tino.pyssysalo@qt.io>