macOS: make creating maintenance tool alias optional
Currently the installer will create an alias file for maintenance tool
with the name defaulting to the file name of maintenance tool, if the
name is not explicitly declared in the installer's configuration file.

This can cause issues with installers that use the same maintenance tool
filename, as the alias would be overwritten on each install, and the
default name does not tell which installation the alias is associated
with. Make it so that the alias is only created if installer config
specifies a name for the alias.

Task-number: QTIFW-2665
Change-Id: I790ee376ec56cff16730189e513a194b8c610066
Reviewed-by: Katja Marttila <katja.marttila@qt.io>