Fix bug when installer is not installing all required components
Several components can have dependencies to a same component. We
calculate the component only once, and increase the reference count to
the dependent component to keep track when it needs to be removed from
install when user clicks the install tree.

The dependency component was removed from install despite of the
reference count. Fixing the bug so that it is truly removed from install
only when there are no references to it.

Task-number: QTIFW-2708
Change-Id: Ieb45270183dc948d9c58271f2d43fccfa7fb8a9f
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>