Replace 'Setup Wizard' and 'Wizard' with just 'Setup'
In many other places the UI refers to the program as just '<Product> Setup'
and the 'Wizard' suffix also seems to be somewhat out of date/'retro'.
So get rid of the suffix to make the naming more consistent.

I think this change also fixes a problem with translation. The ts-files
contain just 'Completing the %1 Setup' for the Restart Page, while the
actual text is 'Completing the %1 Setup Wizard'. So the translation here
might not have worked properly - it certainly looked odd.

Task-Name: SQUISH-9672
Change-Id: I65a56c6e927f01307868ad8319e301b84ab330be
Reviewed-by: Katja Marttila <katja.marttila@qt.io>