Fix bug when installer is not installing required components
Components can be requested to install by several components using
dependency. Components should have a reference count to inform how many
components are requesting the install, so that the component can be
removed from install when nobody has requested it, or user has not
explicitly selected it for install.

Dependency reference count was not calculated correctly, user selection
was ignored if component was already added as dependency. Also if
component which is added as dependency is also autodependency to same
component should not increase the reference count, as the component is
uninstalled when the reference count is zero. Having both autodependency
and dependency would never cause the reference count to be zero.

Task-number: QTIFW-2708
Change-Id: I398bc892f6626d76c872402d8ab55cb8115112ae
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>