Disable package manager and updater for offline maintenance tool
For a maintenance tool that has no configured repositories, we can
disable the selection of package manager and updater options, that
require at least one repository to be useful.

Also show the settings button for offline maintenance tool that is run
in uninstaller mode, because the user may want to add a new repository
which then enables the disabled options.

Task-number: QTIFW-2627
Change-Id: I4db7f6526c37bbadeaf680fe9cb6130058ca5ffd
Reviewed-by: Katja Marttila <katja.marttila@qt.io>