Fix installer stalling when there's only one CPU core
The installer operations are run in the global QThreadPool. The
Operation::performOperation() implementation from "QtPatch" operation
would call the Component::operations() method, that filtered the
returned operation list with QtConcurrent::blockingFilter(). This would
lock the installer when the global thread pool object was initialized
with maxThreadCount of 1, with the only QThread object from the global
pool already running the calling Operation::performOperation() function,
which is waiting the result of Component::operations().

Fix by omitting usage of QtConcurrent::blockingFilter() and populating
the returned operation list with std::copy_if - the components have
usually only a handful operations each so this shouldn't be too
noticeable of a performance regression.

Task-number: QTIFW-2786
Change-Id: Ia6be9f6697310d3f955a8199712c54f345407067
Reviewed-by: Katja Marttila <katja.marttila@qt.io>