Metadatajob: allow replacing existing items to metadata cache
It is a possible situation that a repository with the same Updates.xml
checksum is retrieved from multiple URL:s on a single metadata fetch.
This creates a scenario where more than one metadata item with the same
checksum is pending registration to the cache, which will fail for
the next item after the first registered one. Replacing existing items
should be allowed instead.

Change-Id: I8957202046231d03b3ac26aa8db0534650d49bfb
Reviewed-by: Katja Marttila <katja.marttila@qt.io>