Add possibility to synchronize metadata cache manually
The contents of the cache subdirectories and manifest.json may mismatch
if the installer process is killed after the metadata fetch, because the
manifest file is normally synchronized only when the cache object gets
destroyed. While this is a recoverable error, it creates unnecessary
overhead of discarding broken items and downloading them again.

Add support for synchronizing the cache by the caller and do so each
time the cache is updated by Metadatajob.

Task-number: QTIFW-2817
Change-Id: Ia4cd3de44d57e8d732d11dc26968aa87323ada0f
Reviewed-by: Katja Marttila <katja.marttila@qt.io>