Use static initialization for external-link regexp
Component::updateModelData() is invoked for every value change with any
component, which can result in hundreds of thousands calls when there
is a large number of components. We can avoid the repeated validity
checks for the QRegularExpression object by initializing it as a static
constant.

When refiltering the same repository categories, which omits the
metadata fetching part, there's a time reduction of some 10% from
the start to displaying the component tree again measured by hand,
which would match the results measured with callgrind.

Also add similar handling to the related string literal.

Change-Id: I90694f25e110b54a26ddd0fbc7f9319a6fb5314e
Reviewed-by: Katja Marttila <katja.marttila@qt.io>