Metadata cache: clear cache in a separate thread
This removes the visible blocked UI while clearing cache items, which
takes some time. Disable controls on the dialog and show a basic
progress bar while the clearing is in progress.

Also add basic access serialization to GenericDataCache<T> class
public methods. It is still possible that the user closes the settings
dialog with other means before the clearing in a separate thread was
completed, and we want to avoid crashing to race conditions in such
case. If the user manages to navigate forward in installer pages, to
fetch metadata again, it wouldn't still work correctly though.

Task-number: QTIFW-2815
Change-Id: Ic2c0feac44aff69c458105cf0f559e8693fc0d69
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>
Reviewed-by: Katja Marttila <katja.marttila@qt.io>