ExtractOp: fix leftover empty directories when 'targetDir' arg is used
The problem was that the directory passed for the DirectoryGuard object
handling the creation of leading directories was missing the target
directory itself, instead the path ended to its parent directory.

When installer does the undo-step for Extract operation, only empty
directories are deleted, so in case any directory from a path is
missing from the created directory list, the leading directories
won't be also deleted.

QFileInfo::absolutePath() returns the path of the parent directory for
the given path even if it is a directory, fix by using instead the
QFileInfo::absoluteFilePath(), which includes the name after the
last directory separator.

Task-number: QTIFW-2764
Change-Id: I5b03142b46db566615f4983fa3e2ff2690f25262
Reviewed-by: Katja Marttila <katja.marttila@qt.io>
(cherry picked from commit 5b4a085fb67730f387a9f0228dccc2e47d4e87be)