Workaround possible stalls on single core systems
Try to mitigate risk of running into situations where the application
would hang due to "nested" blocking usage of the global QThreadPool that
has only one thread available. When the ideal thread count for the
machine is 1, which is also the max thread count for the global thread
pool, add another one to the pool.

An example of such hang situation is that the main thread invokes
QtConcurrent::run(&myFilterFunction) and myFilterFunction() calls
QtConcurrent::blockingFiltered() for a container.

Change-Id: Ibe1ee56cb55d7f4fc81c488a381756e09e1c0729
Reviewed-by: Katja Marttila <katja.marttila@qt.io>