MetadataJob: fix removing compressed repositories after extracting
In addition to extracting meta.7z archives, the UnzipArchiveTask is used
to extract compressed repositories (QBSP). At the end of the task, it
removed the archive file uncoditionally, which we don't want to do for
the compressed repositories, because the user should be able to use them
multiple times at their will. Fix by making the removing conditional,
and enable it only for the metadata archives.

Task-number: QTIFW-2876
Change-Id: I10d2fd3872bdcd7eb4bbfeddfd1bda3bdaffeb3f
Reviewed-by: Katja Marttila <katja.marttila@qt.io>