Show check box for AutoDependOn components in updater view
In addition to the automatic selection, allow end user to select
components declaring the AutoDependOn property when updating components.
This does not change the previous rules for automatic selection for
update, meaning the component 'A' declaring AutoDependOn relation to
component 'B' will still be updated when 'B' is selected for update,
regardless of the user selected check state for 'A'.

This fixes some cases not handled by the automatic selection, where
the component declaring AutoDependOn relation could not be updated.

An example scenario:
- 'A' (1.0.0) declaring an auto dependency to 'B' does not follow the
  release schedule and versioning of the component 'B' (2.1.0).
- 'A' has a new release (1.0.1) that is pushed to remote repository
- The end user cannot update the component 'A' unless 'B' is also
  updated to the repository, which would trigger the automatic selection

Task-number: QTIFW-2855
Change-Id: Iea06c366f7f14f391cbc0b4a6526c8aee349ae59
Reviewed-by: Katja Marttila <katja.marttila@qt.io>