Fix freezing UI while searching components
This affected maintenance tool, which by default shows multiple header
columns for the component tree view.

After typing in a search pattern, the component indexes from the tree
are expanded based on the matches. The maintenance tool would visibly
hang at this point. Measuring with callgrind indicated that the biggest
relative cost was at automatically resizing the header sections, which
could be done hundreds of times, which is unnesessary because we only
need to update the view once all necessary indexes are expanded.

Fix by setting the resize modes of the sections temporarily to fixed
size, and restore the original modes after the tree view is updated.

Task-number: QTIFW-2886
Change-Id: I17547344494818de9e321b3501cf6c3bc550c51e
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>
Reviewed-by: Katja Marttila <katja.marttila@qt.io>