Add support for skipping fetching already cached Updates.xml files
This speeds up the metadata evaluation with the Qt online installer,
which uses a dynamic list of temporary repositories that can be
provided with an expected checksum information. The installer or
maintenance tool can skip downloading the Updates.xml files again for
already cached metadata items, in case the checksum of the file in the
repository has not changed since the last metadata fetch.

Task-number: QTIFW-2873
Change-Id: I61384d8bcbd29e01fda3019fb00d7360b0f81435
Reviewed-by: Katja Marttila <katja.marttila@qt.io>