Metadatajob: optimize checking for repository updates
The Updates.xml files from repositories may contain 'RepositoryUpdate'
elements that list actions for existing or new repositories. We need
to parse these updates and restart the metadata job if necessary.
However this was done by parsing the XML file with the Qt XML modules
QDom* classes, which gets expensive as not all of the files contain
any repository updates. This is the most time consuming step left in
case the cache is fully populated and up-to-date.

Instead attempt to make the process faster by filtering out Updates.xml
files that do not contain the 'RepositoryUpdate' element with a string
search, before properly parsing the files.

Change-Id: I87894fdc51f37095e46415a238c9da99624f4f44
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Katja Marttila <katja.marttila@qt.io>