Optimize Updates.xml parsing
QDomDocument::setContent is time consuming operation especially when
there are thousands of Updates.xml to be parsed. Using QXmlStreamReader
for parsing the xml content instead, the Updates.xml content is already
validated in metadata job after download.

In Windows, using QXmlStreamReader instead of QDomDocument decreases the
xml parsing time from ~10 seconds to ~1 seconds when there are thousands
of Updates.xml to be parsed.

Task-number: QTIFW-2924
Change-Id: I14f5eba5b72f451a12ab4e3bf6b2f00b5f66faaf
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>