Tests: fix 'make check' failing after consecutive calls
The 'elevatedexecuteoperation' test invoked qmake in its project file
directory when ran, overwriting the original Makefile. This would cause
problems on the consecutive "make check" invocations after first, if
qmake was originally ran with a custom command line arguments. For
example, setting 'CONFIG+=libarchive' would be lost and the test would
try to link against the default lib7z.a, which would not be found on
libarchive enabled builds.

Fix by calling the qmake executable with the version option instead,
we just need to check that the executable could be started, no need
to actually run it for the test sub-project.

Change-Id: Ie10c1622fa529c1338d427b7c757f8cc5834e0ac
Reviewed-by: Katja Marttila <katja.marttila@qt.io>