Suggest clearing the local cache if loading of meta files fail
In case the UI files or licenses associated with a component cannot be
loaded, one reason may be because the installer was forcibly interrupted
previously and the cache was left in an inconsistent state.

Add a hint about clearing the local cache, so that the files will be
downloaded again as part of the metadata archive for repository (or
component, in case of split metadata format).

Task-number: QTIFW-2883
Change-Id: Ie76365bdc5792a91471cc60037f4375617534fe4
Reviewed-by: Katja Marttila <katja.marttila@qt.io>