Optimize Updates.xml parsing
Updates.xml file elements are read and based on that the installer
calculates for example applicable updates. The files were parsed one by
one which was not very optimal at least in Windows when reading a lot of
files. Fixing the reading so that all files can be read and parsed
using QtConcurrent.

Also changing the class structure so that the private class is removed
and private functions used instead. Another option would have been to
create header file to the private class due to the Q_OBJECT definition
in ParseXMLFilesTask.

Task-number: QTIFW-2805
Change-Id: I40d1c75b87f4bf1f5a8fcd874edf84023891bdcc
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>