Metadata cache: add warning for Updates.xml checksum mismatches
The vendor may provide information about expected checksum for a
repository Updates.xml file, so that installer does not download it
again if it is found in cache. If the expected checksum does not match
the checksum of the actually downloaded file, installer will try to
download it again and again for each metadata fetch.

Detect such mismatch and print a warning to catch those more easily.

Change-Id: I81d2e6e7c4f448229c35ab54c03bcfc3584f304d
Reviewed-by: Katja Marttila <katja.marttila@qt.io>