runextensions: fix forwarding of 'args' parameter pack
The StoredInterface(Member)FunctionCall class templates didn't
initialize their 'args' member variable correctly, it was initialized
with itself. Instead it should be initialized with a tuple variable
created from the forwarded constructor parameters.

Also prefix the member variables of the classes to make it easier to
distinguish the parameter variables of the constructors from the
member ones.

Change-Id: I46acc8f11170b352d2a3fd6e14828d66f13a1cf6
Reviewed-by: Katja Marttila <katja.marttila@qt.io>